Fix another circular dependency between cardutil and Util.
