+++ src/client/utils/Util.ts
@@ -1,9 +1,3 @@
-import Card from 'datatypes/Card';
-import { ColorCategory } from 'datatypes/Card';
-import { TagColor } from 'datatypes/Cube';
-
-import { cardCmc, cardColorIdentityCategory, cardType } from './cardutil';
-
 export function arraysEqual(a: any, b: any): boolean {
   if (a === b) return true;
   if (!Array.isArray(a) || !Array.isArray(b)) return false;
@@ -97,45 +91,6 @@ export function alphaCompare(a: { details: { name: string } }, b: { details: { n
   const textA = a.details.name.toUpperCase();
   const textB = b.details.name.toUpperCase();
   return textA.localeCompare(textB);
-}
-
-export function cmcColumn(card: Card): number {
-  const cmc = cardCmc(card);
-  // Round to half-integer then take ceiling to support Little Girl
-  const cmcDoubleInt = Math.round(cmc * 2);
-  let cmcInt = Math.round((cmcDoubleInt + (cmcDoubleInt % 2)) / 2);
-  if (cmcInt < 0) {
-    cmcInt = 0;
-  }
-  if (cmcInt > 7) {
-    cmcInt = 7;
-  }
-  return cmcInt;
-}
-
-function sortInto(card: Card, result: Card[][][]) {
-  const typeLine = cardType(card).toLowerCase();
-  const row = typeLine.includes('creature') ? 0 : 1;
-  const column = cmcColumn(card);
-  if (result[row][column].length === 0) {
-    result[row][column] = [card];
-  } else {
-    result[row][column].push(card);
-  }
-}
-
-export function sortDeck(deck: (any | any[])[]): any[][] {
-  const result = [new Array(8).fill([]), new Array(8).fill([])];
-  for (const item of deck) {
-    if (Array.isArray(item)) {
-      for (const card of item) {
-        sortInto(card, result);
-      }
-    } else {
-      sortInto(item, result);
-    }
-  }
-  return result;
 }
 
 export const COLORS: [string, string][] = [
@@ -234,36 +189,6 @@ export function isSamePageURL(to: string): boolean {
   }
 }
 
-const colorToColorClass: { [key in ColorCategory]: string } = {
-  White: 'white',
-  Blue: 'blue',
-  Black: 'black',
-  Red: 'red',
-  Green: 'green',
-  Colorless: 'colorless',
-  Multicolored: 'multi',
-  Hybrid: 'multi',
-  Lands: 'lands',
-};
-
-export function getCardColorClass(card: Card): string {
-  if (!card) {
-    return 'colorless';
-  }
-
-  return colorToColorClass[cardColorIdentityCategory(card)];
-}
-
-export function getCardTagColorClass(tagColors: TagColor[], card: Card): string {
-  if (tagColors) {
-    const tagColor = tagColors.find(({ tag }) => (card.tags || []).includes(tag));
-    if (tagColor && tagColor.color && tagColor.color !== 'no-color' && tagColor.color !== 'None') {
-      return `tag-color tag-${tagColor.color}`;
-    }
-  }
-  return getCardColorClass(card);
-}
-
 export function getTagColorClass(tagColors: { tag: string; color: string | null }[], tag: string): string {
   const tagColor = tagColors.find((tagColorB) => tag === tagColorB.tag);
   if (tagColor && tagColor.color && tagColor.color !== 'no-color') {
@@ -311,16 +236,12 @@ export default {
   randomElement,
   fromEntries,
   alphaCompare,
-  cmcColumn,
-  sortDeck,
   COLORS,
   getCubeId,
   getCubeDescription,
   isInternalURL,
   toNullableInt,
   isSamePageURL,
-  getCardColorClass,
-  getCardTagColorClass,
   getTagColorClass,
   wait,
   xorStrings,
